utility to handle user defined point locations
Final version with utility to allow Users to define multiple point location for initial disturbance.
This user input can either be provided directly through command prompt or through text file. In case of file, the text-file should be named as 'point_disturbance.txt' and needs to be placed in the working directory. Information provided in this text file needs to follow an order. First line to reflect number of points provided in the file. Followed by, for every data point, information of x,y coordinate and U values needs to be provided. For example, entry in this text file should look like :5
0.5 0.5 5
0.25 0.25 5
0.75 0.75 5
0.25 0.75 5
0.75 0.25 5Where '5' in first row reflects that data for 5 points are provided. Next line shows data for first point in the form of x y and u value, similarly second line shows data for second point and so on.